Refine Cluster page
