+++ src/Components/Cluster/Cluster.js
@@ -1,9 +1,10 @@
+import './_Cluster.scss';
+
 import React, { useEffect } from 'react';
 import { useIntl } from 'react-intl';
 import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
-import { Card, Grid, GridItem } from '@patternfly/react-core';
-import { List } from 'react-content-loader';
+import { Grid, GridItem } from '@patternfly/react-core';
 
 import PageHeader from '@redhat-cloud-services/frontend-components/PageHeader';
 import Main from '@redhat-cloud-services/frontend-components/Main';
@@ -14,6 +15,7 @@ import ClusterHeader from '../ClusterHeader/ClusterHeader';
 import ClusterRules from '../ClusterRules/ClusterRules';
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
 import { fetchClusterById } from '../../AppActions';
+import Loading from '../Loading/Loading';
 
 const Cluster = ({ match }) => {
   const intl = useIntl();
@@ -35,31 +37,27 @@ const Cluster = ({ match }) => {
 
   return (
     <React.Fragment>
-      {cluster && cluster.fetchStatus === 'rejected' && (
-        <Main>Cluster not found</Main> // TODO: Create empty states
-      )}
-      {cluster && cluster.fetchStatus === 'pending' && (
-        <Card>
-          <List />
-        </Card>
-      )}
-      {cluster && cluster.fetchStatus === 'fulfilled' && (
+      <PageHeader className="pf-m-light ins-inventory-detail">
+        {cluster && (
+          <Breadcrumbs current={match.params.clusterId} match={match} />
+        )}
+        <ClusterHeader uuid={match.params.clusterId} />
+      </PageHeader>
+      <Main>
         <React.Fragment>
-          <PageHeader className="pf-m-light ins-inventory-detail">
-            {cluster && (
-              <Breadcrumbs current={match.params.clusterId} match={match} />
-            )}
-            <ClusterHeader uuid={match.params.clusterId} />
-          </PageHeader>
-          <Main>
+          {cluster && cluster.fetchStatus === 'rejected' && (
+            <React.Fragment>No data for this cluster available</React.Fragment> // TODO: Create empty states
+          )}
+          {cluster && cluster.fetchStatus === 'pending' && <Loading />}
+          {cluster && cluster.fetchStatus === 'fulfilled' && (
             <Grid hasGutter>
               <GridItem span={12}>
                 <ClusterRules cluster={cluster} />
               </GridItem>
             </Grid>
-          </Main>
+          )}
         </React.Fragment>
-      )}
+      </Main>
     </React.Fragment>
   );
 };
